+++ packages/camera/camera_windows/windows/test/capture_controller_test.cpp
@@ -33,7 +33,8 @@ void MockInitCaptureController(
     CaptureControllerImpl* capture_controller,
     MockTextureRegistrar* texture_registrar, MockCaptureEngine* engine,
     MockCamera* camera, int64_t mock_texture_id,
-    const RecordSettings record_settings = RecordSettings(true)) {
+    const PlatformMediaSettings media_settings =
+        PlatformMediaSettings(PlatformResolutionPreset::max, true)) {
   ComPtr<MockMediaSource> video_source = new MockMediaSource();
   ComPtr<MockMediaSource> audio_source = new MockMediaSource();
 
@@ -61,8 +62,7 @@ void MockInitCaptureController(
   EXPECT_CALL(*engine, Initialize).Times(1);
 
   bool result = capture_controller->InitCaptureDevice(
-      texture_registrar, MOCK_DEVICE_ID, ResolutionPreset::kAuto,
-      record_settings);
+      texture_registrar, MOCK_DEVICE_ID, media_settings);
 
   EXPECT_TRUE(result);
 
@@ -260,8 +260,8 @@ TEST(CaptureController, InitCaptureEngineCanOnlyBeCalledOnce) {
   EXPECT_CALL(*camera, OnCreateCaptureEngineFailed).Times(1);
 
   bool result = capture_controller->InitCaptureDevice(
-      texture_registrar.get(), MOCK_DEVICE_ID, ResolutionPreset::kAuto,
-      RecordSettings(true));
+      texture_registrar.get(), MOCK_DEVICE_ID,
+      PlatformMediaSettings(PlatformResolutionPreset::max, true));
 
   EXPECT_FALSE(result);
 
@@ -302,8 +302,8 @@ TEST(CaptureController, InitCaptureEngineReportsFailure) {
       .Times(1);
 
   bool result = capture_controller->InitCaptureDevice(
-      texture_registrar.get(), MOCK_DEVICE_ID, ResolutionPreset::kAuto,
-      RecordSettings(true));
+      texture_registrar.get(), MOCK_DEVICE_ID,
+      PlatformMediaSettings(PlatformResolutionPreset::max, true));
 
   EXPECT_FALSE(result);
   EXPECT_FALSE(engine->initialized_);
@@ -347,8 +347,8 @@ TEST(CaptureController, InitCaptureEngineReportsAccessDenied) {
       .Times(1);
 
   bool result = capture_controller->InitCaptureDevice(
-      texture_registrar.get(), MOCK_DEVICE_ID, ResolutionPreset::kAuto,
-      RecordSettings(true));
+      texture_registrar.get(), MOCK_DEVICE_ID,
+      PlatformMediaSettings(PlatformResolutionPreset::max, true));
 
   EXPECT_FALSE(result);
   EXPECT_FALSE(engine->initialized_);
@@ -734,16 +734,15 @@ TEST(CaptureController, StartRecordWithSettingsSuccess) {
   const auto kVideoBitrate = 200000;
   const auto kAudioBitrate = 32000;
 
-  RecordSettings record_settings;
-  record_settings.record_audio = true;
-  record_settings.fps = kFps;
-  record_settings.video_bitrate = kVideoBitrate;
-  record_settings.audio_bitrate = kAudioBitrate;
+  PlatformMediaSettings media_settings(PlatformResolutionPreset::max, true);
+  media_settings.set_frames_per_second(kFps);
+  media_settings.set_video_bitrate(kVideoBitrate);
+  media_settings.set_audio_bitrate(kAudioBitrate);
 
   // Initialize capture controller to be able to start preview
   MockInitCaptureController(capture_controller.get(), texture_registrar.get(),
                             engine.Get(), camera.get(), mock_texture_id,
-                            record_settings);
+                            media_settings);
 
   ComPtr<MockCaptureSource> capture_source = new MockCaptureSource();