+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/MarkersController.java
@@ -10,7 +10,7 @@ import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.MarkerOptions;
 import com.google.maps.android.collections.MarkerManager;
-import io.flutter.plugin.common.MethodChannel;
+import io.flutter.plugins.googlemaps.Messages.MapsCallbackApi;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -20,21 +20,21 @@ class MarkersController {
   private final HashMap<String, MarkerBuilder> markerIdToMarkerBuilder;
   private final HashMap<String, MarkerController> markerIdToController;
   private final HashMap<String, String> googleMapsMarkerIdToDartMarkerId;
-  private final MethodChannel methodChannel;
+  private final @NonNull MapsCallbackApi flutterApi;
   private MarkerManager.Collection markerCollection;
   private final ClusterManagersController clusterManagersController;
   private final AssetManager assetManager;
   private final float density;
 
   MarkersController(
-      MethodChannel methodChannel,
+      @NonNull MapsCallbackApi flutterApi,
       ClusterManagersController clusterManagersController,
       AssetManager assetManager,
       float density) {
     this.markerIdToMarkerBuilder = new HashMap<>();
     this.markerIdToController = new HashMap<>();
     this.googleMapsMarkerIdToDartMarkerId = new HashMap<>();
-    this.methodChannel = methodChannel;
+    this.flutterApi = flutterApi;
     this.clusterManagersController = clusterManagersController;
     this.assetManager = assetManager;
     this.density = density;
@@ -47,7 +47,9 @@ class MarkersController {
   void addJsonMarkers(List<Object> markersToAdd) {
     if (markersToAdd != null) {
       for (Object markerToAdd : markersToAdd) {
-        addJsonMarker(markerToAdd);
+        @SuppressWarnings("unchecked")
+        Map<String, ?> markerMap = (Map<String, ?>) markerToAdd;
+        addJsonMarker(markerMap);
       }
     }
   }
@@ -126,7 +128,7 @@ class MarkersController {
   }
 
   boolean onMarkerTap(String markerId) {
-    methodChannel.invokeMethod("marker#onTap", Convert.markerIdToJson(markerId));
+    flutterApi.onMarkerTap(markerId, new NoOpVoidResult());
     MarkerController markerController = markerIdToController.get(markerId);
     if (markerController != null) {
       return markerController.consumeTapEvents();
@@ -139,10 +141,7 @@ class MarkersController {
     if (markerId == null) {
       return;
     }
-    final Map<String, Object> data = new HashMap<>();
-    data.put("markerId", markerId);
-    data.put("position", Convert.latLngToJson(latLng));
-    methodChannel.invokeMethod("marker#onDragStart", data);
+    flutterApi.onMarkerDragStart(markerId, Convert.latLngToPigeon(latLng), new NoOpVoidResult());
   }
 
   void onMarkerDrag(String googleMarkerId, LatLng latLng) {
@@ -150,10 +149,7 @@ class MarkersController {
     if (markerId == null) {
       return;
     }
-    final Map<String, Object> data = new HashMap<>();
-    data.put("markerId", markerId);
-    data.put("position", Convert.latLngToJson(latLng));
-    methodChannel.invokeMethod("marker#onDrag", data);
+    flutterApi.onMarkerDrag(markerId, Convert.latLngToPigeon(latLng), new NoOpVoidResult());
   }
 
   void onMarkerDragEnd(String googleMarkerId, LatLng latLng) {
@@ -161,10 +157,7 @@ class MarkersController {
     if (markerId == null) {
       return;
     }
-    final Map<String, Object> data = new HashMap<>();
-    data.put("markerId", markerId);
-    data.put("position", Convert.latLngToJson(latLng));
-    methodChannel.invokeMethod("marker#onDragEnd", data);
+    flutterApi.onMarkerDragEnd(markerId, Convert.latLngToPigeon(latLng), new NoOpVoidResult());
   }
 
   void onInfoWindowTap(String googleMarkerId) {
@@ -172,7 +165,7 @@ class MarkersController {
     if (markerId == null) {
       return;
     }
-    methodChannel.invokeMethod("infoWindow#onTap", Convert.markerIdToJson(markerId));
+    flutterApi.onInfoWindowTap(markerId, new NoOpVoidResult());
   }
 
   /**
@@ -186,7 +179,7 @@ class MarkersController {
     }
   }
 
-  private void addJsonMarker(Object marker) {
+  private void addJsonMarker(Map<String, ?> marker) {
     if (marker == null) {
       return;
     }
@@ -232,7 +225,7 @@ class MarkersController {
     googleMapsMarkerIdToDartMarkerId.put(marker.getId(), markerId);
   }
 
-  private void changeJsonMarker(Object marker) {
+  private void changeJsonMarker(Map<String, ?> marker) {
     if (marker == null) {
       return;
     }
@@ -264,15 +257,11 @@ class MarkersController {
     }
   }
 
-  @SuppressWarnings("unchecked")
-  private static String getMarkerId(Object marker) {
-    Map<String, Object> markerMap = (Map<String, Object>) marker;
-    return (String) markerMap.get("markerId");
+  private static String getMarkerId(Map<String, ?> marker) {
+    return (String) marker.get("markerId");
   }
 
-  @SuppressWarnings("unchecked")
-  private static String getClusterManagerId(Object marker) {
-    Map<String, Object> markerMap = (Map<String, Object>) marker;
-    return (String) markerMap.get("clusterManagerId");
+  private static String getClusterManagerId(Map<String, ?> marker) {
+    return (String) marker.get("clusterManagerId");
   }
 }