+++ packages/google_maps_flutter/google_maps_flutter_ios/ios/Classes/FLTGoogleMapJSONConversions.m
@@ -10,6 +10,28 @@ id FGMGetValueOrNilFromDict(NSDictionary *dict, NSString *key) {
   return value == [NSNull null] ? nil : value;
 }
 
+CGPoint FGMGetCGPointForPigeonPoint(FGMPlatformPoint *point) {
+  return CGPointMake(point.x, point.y);
+}
+
+FGMPlatformPoint *FGMGetPigeonPointForCGPoint(CGPoint point) {
+  return [FGMPlatformPoint makeWithX:point.x y:point.y];
+}
+
+CLLocationCoordinate2D FGMGetCoordinateForPigeonLatLng(FGMPlatformLatLng *latLng) {
+  return CLLocationCoordinate2DMake(latLng.latitude, latLng.longitude);
+}
+
+FGMPlatformLatLng *FGMGetPigeonLatLngForCoordinate(CLLocationCoordinate2D coord) {
+  return [FGMPlatformLatLng makeWithLatitude:coord.latitude longitude:coord.longitude];
+}
+
+FGMPlatformLatLngBounds *FGMGetPigeonLatLngBoundsForCoordinateBounds(GMSCoordinateBounds *bounds) {
+  return
+      [FGMPlatformLatLngBounds makeWithNortheast:FGMGetPigeonLatLngForCoordinate(bounds.northEast)
+                                       southwest:FGMGetPigeonLatLngForCoordinate(bounds.southWest)];
+}
+
 @implementation FLTGoogleMapJSONConversions
 
 + (CLLocationCoordinate2D)locationFromLatLong:(NSArray *)latlong {
@@ -67,23 +89,6 @@ id FGMGetValueOrNilFromDict(NSDictionary *dict, NSString *key) {
   };
 }
 
-+ (NSDictionary<NSString *, NSNumber *> *)dictionaryFromPoint:(CGPoint)point {
-  return @{
-    @"x" : @(lroundf(point.x)),
-    @"y" : @(lroundf(point.y)),
-  };
-}
-
-+ (nullable NSDictionary *)dictionaryFromCoordinateBounds:(GMSCoordinateBounds *)bounds {
-  if (!bounds) {
-    return nil;
-  }
-  return @{
-    @"southwest" : [FLTGoogleMapJSONConversions arrayFromLocation:[bounds southWest]],
-    @"northeast" : [FLTGoogleMapJSONConversions arrayFromLocation:[bounds northEast]],
-  };
-}
-
 + (nullable GMSCameraPosition *)cameraPostionFromDictionary:(nullable NSDictionary *)data {
   if (!data) {
     return nil;
@@ -93,12 +98,6 @@ id FGMGetValueOrNilFromDict(NSDictionary *dict, NSString *key) {
                   zoom:[data[@"zoom"] floatValue]
                bearing:[data[@"bearing"] doubleValue]
           viewingAngle:[data[@"tilt"] doubleValue]];
-}
-
-+ (CGPoint)pointFromDictionary:(NSDictionary *)dictionary {
-  double x = [dictionary[@"x"] doubleValue];
-  double y = [dictionary[@"y"] doubleValue];
-  return CGPointMake(x, y);
 }
 
 + (GMSCoordinateBounds *)coordinateBoundsFromLatLongs:(NSArray *)latlongs {
@@ -112,7 +111,7 @@ id FGMGetValueOrNilFromDict(NSDictionary *dict, NSString *key) {
   return (GMSMapViewType)(value == 0 ? 5 : value);
 }
 
-+ (nullable GMSCameraUpdate *)cameraUpdateFromChannelValue:(NSArray *)channelValue {
++ (nullable GMSCameraUpdate *)cameraUpdateFromArray:(NSArray *)channelValue {
   NSString *update = channelValue[0];
   if ([update isEqualToString:@"newCameraPosition"]) {
     return [GMSCameraUpdate