+++ packages/google_maps_flutter/google_maps_flutter_ios/ios/Classes/FLTGoogleMapTileOverlayController.h
@@ -5,15 +5,19 @@
 #import <Flutter/Flutter.h>
 #import <GoogleMaps/GoogleMaps.h>
 
+#import "messages.g.h"
+
 NS_ASSUME_NONNULL_BEGIN
 
 @interface FLTGoogleMapTileOverlayController : NSObject
+/// The layer managed by this controller instance.
+@property(readonly, nonatomic) GMSTileLayer *layer;
+
 - (instancetype)initWithTileLayer:(GMSTileLayer *)tileLayer
                           mapView:(GMSMapView *)mapView
                           options:(NSDictionary *)optionsData;
 - (void)removeTileOverlay;
 - (void)clearTileCache;
-- (NSDictionary *)getTileOverlayInfo;
 @end
 
 @interface FLTTileProviderController : GMSTileLayer
@@ -26,11 +30,12 @@ NS_ASSUME_NONNULL_BEGIN
 - (instancetype)init:(FlutterMethodChannel *)methodChannel
              mapView:(GMSMapView *)mapView
            registrar:(NSObject<FlutterPluginRegistrar> *)registrar;
-- (void)addTileOverlays:(NSArray *)tileOverlaysToAdd;
-- (void)changeTileOverlays:(NSArray *)tileOverlaysToChange;
-- (void)removeTileOverlayWithIdentifiers:(NSArray *)identifiers;
+- (void)addJSONTileOverlays:(NSArray<NSDictionary<NSString *, id> *> *)tileOverlaysToAdd;
+- (void)addTileOverlays:(NSArray<FGMPlatformTileOverlay *> *)tileOverlaysToAdd;
+- (void)changeTileOverlays:(NSArray<FGMPlatformTileOverlay *> *)tileOverlaysToChange;
+- (void)removeTileOverlayWithIdentifiers:(NSArray<NSString *> *)identifiers;
 - (void)clearTileCacheWithIdentifier:(NSString *)identifier;
-- (nullable NSDictionary *)tileOverlayInfoWithIdentifier:(NSString *)identifier;
+- (nullable FLTGoogleMapTileOverlayController *)tileOverlayWithIdentifier:(NSString *)identifier;
 @end
 
 NS_ASSUME_NONNULL_END